chore: Removed hot reloading feature.
