+++ package.json
@@ -29,23 +29,19 @@
   "main": "dist/main.js",
   "scripts": {
     "start": "run-s clean build serve",
-    "predev": "rimraf dist",
-    "dev": "run-p dev:*",
-    "dev:electron": "wait-on ./dist/index.html && cross-env NODE_ENV=\"development\" electron .",
-    "dev:tsc": "tsc -w -p tsconfig.main.json",
-    "dev:webpack": "webpack --watch",
     "serve": "electron .",
     "clean": "rimraf dist release",
-    "build": "webpack --config webpack.config.prod.ts --progress",
+    "build": "cross-env NODE_ENV=\"development\" webpack --progress",
+    "build:prod": "cross-env NODE_ENV=\"production\" webpack --progress",
     "build:unsign": "ts-node scripts/build-unsign.ts",
     "build:appx": "ts-node scripts/build-appx.ts",
     "build:darwin": "ts-node scripts/build.ts",
     "lint": "run-s lint:*",
     "lint:eslint": "eslint . --ext .ts,.tsx --fix",
     "lint:prettier": "prettier --write .",
-    "package": "run-s clean build build:unsign",
-    "package:appx": "run-s clean build build:appx",
-    "package:darwin": "run-s clean build build:darwin",
+    "package": "run-s clean build:prod build:unsign",
+    "package:appx": "run-s clean build:prod build:appx",
+    "package:darwin": "run-s clean build:prod build:darwin",
     "prepare": "git config --local core.hooksPath .githooks"
   },
   "dependencies": {
@@ -79,7 +75,6 @@
     "electron-builder": "^22.11.7",
     "electron-notarize": "^1.1.1",
     "html-webpack-plugin": "^5.3.2",
-    "ifdef-loader": "^2.3.0",
     "mini-css-extract-plugin": "^2.3.0",
     "npm-run-all": "^4.1.5",
     "rimraf": "^3.0.2",
@@ -94,13 +89,11 @@
   "optionalDependencies": {
     "@typescript-eslint/eslint-plugin": "^4.31.1",
     "@typescript-eslint/parser": "^4.31.1",
-    "electron-reload": "^1.5.0",
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-react": "^7.25.1",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "prettier": "^2.4.0",
-    "wait-on": "^6.0.0"
+    "prettier": "^2.4.0"
   },
   "resolutions": {
     "electron-builder/dmg-builder/dmg-license/plist/xmldom": "github:xmldom/xmldom"
